requirements.txt
brainflow
h5py
joblib
matplotlib 
mne
natsort
numpy
pandas
pygame
pytest
python-dotenv
pyzmq
scipy
seaborn
sklearn
